Project structure update
