+++ nameless/commons/Utility.py
@@ -63,10 +63,10 @@ class Utility:
         finish_url = (
             f"{dialect}{pdriver}://{username}{ppassword}{at}{host}{pport}/{db_name}"
         )
-        secret_url = (
+        hidden_url = (
             f"{dialect}{pdriver}://{username}{vpassword}{at}{vhost}{vport}/{db_name}"
         )
-        logging.info("Using %s as database URL", secret_url)
+        logging.info("Using %s as database URL", hidden_url)
 
         return (
             finish_url,
