+++ nameless/customs/ColoredFormatter.py
