+++ nameless/customs/__init__.py
@@ -1,3 +1,2 @@
 from .ColoredFormatter import *
-from .Utility import *
 from .DiscordWaiter import *
