+++ tests/test_util.py
@@ -1,4 +1,4 @@
-from customs import Utility
+from nameless.customs import Utility
 from config_example import Config
